+++ packages/terra-site/CHANGELOG.md
@@ -3,6 +3,10 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Add terra-form-textarea components
+* Add terra-form-field components
+* Add terra-mock-theme styles
 
 1.17.0 - (November 7, 2017)
 ------------------